May the forrce be with you

+++ ingsoc.c
@@ -1,6 +1,6 @@
 
 #include "ingsoc.h"
-
+#include <sodium.h>
 
 void ingsoc_init(ingsoc *insoci)
 {
@@ -18,7 +18,7 @@ void ingsoc_init(ingsoc *insoci)
 
 void ingsoc_seqnr(ingsoc *in)
 {
-    static size_t startNr=0;
+    static uint32_t startNr=0;
     if(startNr == 0){
         startNr = randombytes_uniform(2147483647) + 10;
     }else{
